purge: do not delete contents of oci indexes
The purge command attempts to detect if manifest lists are being used,
and will avoid deleting manifests that are part of them.

The OCI specification defines an index media type, that can contain
manifests as well, as an alternative to manifest lists - functionally,
these are the same as manifest lists, and ACR appears to support them.

The purge command should therefore avoid removing manifests that are
part of indexes as well, which, if the user is using OCI images, will
cause their images to be unintentionally deleted.

Signed-off-by: Justin Chadwell <me@jedevc.com>